Filter removal. (#1536)
* Filter removal.

* Fix check.

+++ src/UniversalDashboard/Services/FilterProvider.cs
@@ -20,11 +20,15 @@ namespace UniversalDashboard
 
         public void OnProvidersExecuting(FilterProviderContext context)
         {
-            // remove authorize filters
-            var authFilters = context.Results.Where(x => 
-            x.Descriptor.Filter.GetType() == typeof(AuthorizeFilter)).ToList();
-            foreach(var f in authFilters)
-                context.Results.Remove(f);
+            var filterRemoval = context.ActionContext.HttpContext.RequestServices.GetService(typeof(IFilterRemoval)) as IFilterRemoval;
+            if (filterRemoval == null || filterRemoval.ShouldRemove(context)) 
+            {
+                // remove authorize filters
+                var authFilters = context.Results.Where(x => 
+                x.Descriptor.Filter.GetType() == typeof(AuthorizeFilter)).ToList();
+                foreach(var f in authFilters)
+                    context.Results.Remove(f);
+            }
         }
     }
 }
